feat: complete refactor of java classes.
