+++ Software/LMSourceCode/ImageProcessing/golfsim_tomee_webapp/src/main/java/com/verdanttechs/jakarta/ee9/enums/GsClubType.java
@@ -1,0 +1,8 @@
+package com.verdanttechs.jakarta.ee9.enums;
+
+public enum GsClubType {
+    kNotSelected,
+    kDriver,
+    kIron,
+    kPutter;
+}
