+++ Software/LMSourceCode/ImageProcessing/golfsim_tomee_webapp/src/main/java/com/verdanttechs/jakarta/ee9/enums/GsIPCResultType.java
@@ -1,0 +1,14 @@
+package com.verdanttechs.jakarta.ee9.enums;
+
+public enum GsIPCResultType {
+        kUnknown,
+        kInitializing,
+        kWaitingForBallToAppear,
+        kWaitingForSimulatorArmed,
+        kPausingForBallStabilization,
+        kMultipleBallsPresent,
+        kBallPlacedAndReadyForHit,
+        kHit,
+        kError,
+        kCalibrationResults;
+    }
