+++ Software/LMSourceCode/ImageProcessing/golfsim_tomee_webapp/src/main/java/com/verdanttechs/jakarta/ee9/enums/IPCMessageType.java
@@ -1,0 +1,12 @@
+package com.verdanttechs.jakarta.ee9.enums;
+
+public enum IPCMessageType {
+    kUnknown,
+    kRequestForCamera2Image,
+    kCamera2Image,
+    kRequestForCamera2TestStillImage,
+    kResults,
+    kShutdown,
+    kCamera2ReturnPreImage,
+    kControlMessage;
+}
